\name{COMPoisson-internal}

\alias{COMP_dnu_objfn}
\alias{COMP_maxn}
\alias{COMP_Z}
\alias{COMP_Z_lfacn}
\alias{COMP_Z_n}
\alias{COMP_Z_n2}
\alias{COMP_Z_n3}
\alias{COMP_Z_ratio}
\alias{dCOMP}
\alias{Rcpp_COMP_Z}

\title{Internal Functions for COM-Poisson family}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
