\name{spaMM-internal-blockDiag}

\alias{RcppChol.blockDiag} 
\alias{as.blockDiag.bar}
\alias{as.blockDiag.formula}
\alias{as.blockDiag.list}
\alias{as.blockDiag.matrix} 
\alias{as.blockDiag.partition} 
\alias{as.matrix.blockDiag}
\alias{chol.blockDiag} 
\alias{findblocks}
\alias{t.blockDiag}

\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
