\name{spaMM-internal-migraineLike}

\alias{elim.redundant.V} 
\alias{rsimplex}
\alias{upperPoints}

\alias{connectedSets}
\alias{extrapolhull}
%~ 'old':
\alias{old.nextPoints} 
\alias{ordinaryKrigingCoefs}
\alias{rtnorm.copy}
\alias{sampleNearby}
\alias{sampleNextPoints}

\title{Internal spaMM Functions}
\description{
  Internal spaMM functions, ultimately for smoothing as in Migraine
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }


