\name{spaMM-Rcpp-internal}

\alias{Cholwrap}
\alias{crossprodCpp} 
\alias{lmwithQ} 
\alias{lmwithSparseQ} 
\alias{LevPerturbedQCpp}
\alias{LevenbergMsolveCpp}
\alias{leverages}
\alias{LogAbsDetCpp}
\alias{pseudoSolvediag}
\alias{QRwrap}
\alias{Rcpp_QR}
\alias{Rcpp_sparseQR}
\alias{Rcpp_qr_Q} 
\alias{RcppChol}
\alias{selfAdjointSolverCpp} 
\alias{sweepZ1W} 
\alias{tcrossprodCpp} 
\alias{ZWZt} 
\alias{ZtWZ} 
\alias{Rcpp_pMVN}

\title{Internal spaMM Functions using Rcpp/RcppEigen}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
