\name{spaMM-Rcpp-internal}

%\alias{Cholwrap}
%\alias{crossprodCpp} 
%\alias{LevPerturbedQCpp}
%\alias{LevenbergMsolve_Matrix}
%\alias{LevenbergMsolveCpp}
%\alias{e_LevenbergMsolveCpp}
%\alias{leverages}
%\alias{LogAbsDetCpp}
%\alias{pseudoSolvediag}
%\alias{QRwrap}
%\alias{Rcpp_Sig}
%\alias{Rcpp_d2hdv2}
%\alias{RcppChol}
%\alias{sweepZ1W} 
%\alias{tcrossprodCpp} 
%\alias{ZWZt} % used by probitgem !
%\alias{ZtWZ} 
%\alias{Rcpp_pMVN}
%\alias{zrnorm}

\title{Internal spaMM Functions using Rcpp/RcppEigen}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
