\name{spaMM-internal}


\alias{alternating}
\alias{Bartlett.robust} 
\alias{Beta} 
\alias{checkLink}
\alias{checkLinkS}
\alias{compute.ZALlist} 
\alias{CondNormfn} 
\alias{Dim.corMatern} 
\alias{GammaForDispGammaGLM}
\alias{HLCor.obj} 
\alias{HL_process.args} 
\alias{HL_strictify} 
\alias{HLframes}
\alias{Initialize.corMatern} 
\alias{LR.HL} 
\alias{LRT.corrHL} 
\alias{Matern.local} 
\alias{ULI}
\alias{Variogram.corMatern} 
\alias{asStandardFormula} 
\alias{bessel_lnKnu} 
\alias{calc.tXinvS}
\alias{calcDhDv2}
\alias{calcPredVar}
\alias{clean_cs} 
\alias{checkDistMatrix} 
\alias{coef.corMatern} 
\alias{coef<-.corMatern} 
\alias{corFactor.corMatern}
\alias{corMatrix.corMatern} 
\alias{corrMM.LRT} 
\alias{d2mudeta2fn} 
\alias{DEPARSE}
\alias{dispFn}
\alias{dispGammaGLM} 
\alias{dispInv} 
\alias{extract.check.coords} 
\alias{findOffset} 
\alias{findSpatial} 
\alias{findSpatialOrNot} 
\alias{findbarsMM} 
\alias{findplus}
\alias{formula.HLfit}
\alias{generateFileName} 
\alias{generateName} 
\alias{getCovariate.corMatern} 
\alias{getgroups} 
\alias{is.square.diagonal} 
\alias{is.identity} 
\alias{largeLambdaMessages} 
\alias{legend_lambda}
\alias{LevenbergMstep}
\alias{LevenbergMstepCallingCpp}
\alias{locoptim}
\alias{locoptimthroughSmooth}
\alias{logDet.corMatern} 
\alias{loglfn.ranU} 
\alias{makeCheckGeoMatrices}
\alias{makeCovEst} 
\alias{muetafn}
\alias{newetaFix}
\alias{nextPoints}
\alias{nobarsMM} 
\alias{nobarsNooffset} 
\alias{noOffset} 
\alias{noReForm} 
\alias{nuFn} 
\alias{nuInv} 
\alias{pastefrom} 
\alias{plotdhglm} 
\alias{post.process.ZALlist}
\alias{preprocess}
%- \alias{predict.HL}
\alias{print.arglist}
\alias{print.predictor}
\alias{recalc.corMatern} 
\alias{rhoFn} 
\alias{rhoInv}
\alias{rntneg} 
\alias{rntpos} 
\alias{safesolve.qr.matrix} 
\alias{safesolve.qr.vector} 
\alias{sampleGridFromLowUp}
\alias{selectLoglfn}
\alias{separator} 
\alias{slashTerms} 
\alias{solveWrap.matrix}
\alias{solveWrap.vector}
\alias{spaMMLRT} 
\alias{spMMFactorList} 
\alias{spMMexpandSlash}
\alias{subbarsMM} 
\alias{theta.mu.canonical}
\alias{theta.mu.conjugate} 
\alias{thetaMuDerivs} 
\alias{tryCatch.W.E}
\alias{updateWranef}
\alias{updateW_ranefS}
\alias{validRows}
\alias{volTriangulation} %~ could become public later : cf le .Rd.true.txt
\alias{rvolTriangulation} %~ could become public later




\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
