\name{spaMM-internal}

\alias{\%id*id\%}
\alias{alternating}
\alias{auglinmodfit}
\alias{Bartlett.robust} 
\alias{Beta} 
\alias{calc.p_v}
\alias{calc.w.resid}
\alias{calc.hatval}
\alias{calc.Pdiag}
\alias{calc.plot.dims}
\alias{calcDhDv2}
\alias{CBIND}
\alias{CBIND.matrix}
\alias{CBIND.Matrix}
\alias{checkRandLinkS}
\alias{checkRandLink}
\alias{checkRespFam}
\alias{compare.model.structures}
\alias{compute.ZALlist} 
\alias{CondNormfn} 
\alias{corr.notEQL.lambda}
\alias{Dim.corMatern} 
\alias{eval.update.call}
\alias{eval.corrEst.args}
\alias{GammaForDispGammaGLM}
\alias{getCallHL}
\alias{getHLfit}
\alias{HLCor.obj} 
\alias{HL_process.args} 
\alias{HL_strictify} 
\alias{HLframes}
\alias{Initialize.corMatern}
\alias{intervalStep}
\alias{LRT.corrHL} 
\alias{loglfn.ranU}
\alias{makenewname}
\alias{Matern.local} 
\alias{ULI}
\alias{Variogram.corMatern} 
\alias{asStandardFormula} 
\alias{bessel_lnKnu} 
\alias{calc.tXinvS}
\alias{calcDhDv2}
\alias{calcNewCorrs}
\alias{calcPredVar}
\alias{calcResidVar}
\alias{canonizeRanPars} 
\alias{clean_cs} 
\alias{checkDistMatrix} 
\alias{coef.corMatern} 
\alias{coef<-.corMatern} 
\alias{corFactor.corMatern}
\alias{corMatrix.corMatern} 
\alias{corrMM.LRT} 
\alias{d2mudeta2fn} 
\alias{DEPARSE}
\alias{dispFn}
\alias{dispGammaGLM} 
\alias{dispInv} 
\alias{extract.check.coords} 
\alias{findOffset} 
\alias{findSpatial} 
\alias{parseBars} 
\alias{findbarsMM} 
\alias{findplus}
\alias{formula.HLfit}
\alias{generateFileName} 
\alias{generateName} 
\alias{getCovariate.corMatern} 
\alias{getgroups} 
\alias{getPar}
\alias{getProcessed}
\alias{is.square.diagonal} 
\alias{is.identity} 
\alias{largeLambdaMessages} 
\alias{legend_lambda}
\alias{LevenbergMstep}
\alias{LevenbergMstepCallingCpp}
\alias{locoptim}
\alias{locoptimthroughSmooth}
\alias{logDet.corMatern} 
\alias{LogAbsDetWrap}
\alias{loglfn.ranU} 
\alias{makeCheckGeoMatrices}
\alias{makeCovEst2} 
\alias{makeLowerUpper}
\alias{muetafn}
\alias{newetaFix}
\alias{nextPoints}
\alias{nobarsMM} 
\alias{nobarsNooffset} 
\alias{noOffset} 
\alias{noReForm} 
\alias{nuFn} 
\alias{nuInv} 
\alias{pastefrom} 
\alias{plotdhglm} 
\alias{plotScale}
\alias{post.process.ZALlist}
\alias{preprocess}
\alias{print.arglist}
\alias{print.predictor}
\alias{print.ranef}
\alias{RBIND}
\alias{RBIND.matrix}
\alias{RBIND.Matrix}
\alias{recalc.corMatern} 
\alias{recent.mapMM} %~ should disappear
\alias{rhoFn} 
\alias{rhoInv}
\alias{rntneg} 
\alias{rntpos} 
\alias{safesolve.qr.matrix} 
\alias{safesolve.qr.vector} 
\alias{sampleGridFromLowUp}
\alias{selectLoglfn}
\alias{separator} 
\alias{setProcessed}
\alias{Sigwrapper}
\alias{slashTerms} 
\alias{solveWrap.matrix}
\alias{solveWrap.vector}
\alias{spaMMLRT} 
\alias{spMMFactorList} 
\alias{spMMexpandSlash}
\alias{subbarsMM} 
\alias{theta.mu.canonical}
\alias{theta.mu.conjugate} 
\alias{thetaMuDerivs} 
\alias{tryCatch.W.E}
\alias{u_h_from_v_h}
\alias{updateWranef}
\alias{updateW_ranefS}
\alias{validData}
\alias{ZWZtwrapper}
\alias{ZtWZwrapper}
\alias{volTriangulation} %~ could become public later : cf le .Rd.true.txt
\alias{rvolTriangulation} %~ could become public later
\alias{subsimplices.volTriangulation} %~ could become public later




\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
