
\name{spaMM-internal}

\alias{fitme_body}
\alias{calc_maxrange}
\alias{calc_inits_dispPars}
\alias{calc_inits_ARphi}
\alias{calc_inits_nu}
\alias{calc_inits_Auto_rho}
\alias{calc_inits_geostat_rho}
\alias{calc_inits}
\alias{HLfit.obj}




\alias{\%id*id\%}
\alias{alternating}
\alias{auglinmodfit}
\alias{Bartlett.robust} 
\alias{Beta} 
\alias{bloc_lambda}
\alias{calc_beta_w_cov}
\alias{calc.dlW_deta}
\alias{calc.dvdloglamMat}
\alias{calc_invS.X}
\alias{calc.p_v}
\alias{calc.w.resid}
\alias{calc_wAugX}
\alias{calc_hatval}
\alias{calc.Pdiag}
\alias{calc.plot.dims}
\alias{calcD2hDv2}
\alias{calcPHI}
\alias{calc_dispGammaGLM}
\alias{calcRanefPars}
\alias{calcTT}
\alias{calcUniqueGeo}
\alias{calc_info_crits}
\alias{calc_invL}
\alias{calc_invL_coeffs}
\alias{CBIND}
\alias{CBIND.matrix}
\alias{CBIND.Matrix}
\alias{checkRandLinkS}
\alias{checkRandLink}
\alias{checkRespFam}
\alias{compare.model.structures}
\alias{computeZAXlist} 
\alias{CondNormfn} 
\alias{corr.notEQL.lambda}
\alias{create_get_w_h_coeffs}
\alias{create_get_beta_w_cov}
\alias{create_get_invColdoldList}
\alias{create_get_logdispObject}
\alias{create_get_info_crits}
\alias{create_get_init_phi}
\alias{Dim.corMatern} 
\alias{eval.update.call}
\alias{eval.corrEst.args}
\alias{generateInitLambda}
\alias{spaMM_Gamma} 
\alias{getHLfit}
\alias{get_valid_beta_coefs}
\alias{HLCor.obj} 
\alias{HL_process.args} 
\alias{HL_strictify} 
\alias{HLframes}
\alias{Initialize.corMatern}
\alias{initialize.v_h}
\alias{intervalStep}
\alias{iterateSEMSmooth}
%\alias{locatePointinvT}
\alias{LRT.corrHL} 
\alias{loglfn.ranU}
\alias{makenewname}
\alias{ULI}
\alias{Variogram.corMatern} 
\alias{asStandardFormula} 
\alias{bessel_lnKnu} 
\alias{calc_tXinvS}
\alias{calcNewCorrs}
\alias{calcPredVar}
\alias{calcResidVar}
\alias{calc_beta_cov}
\alias{calc_dvdlogphiMat}
\alias{calc_invColdoldList}
\alias{calc_logdisp_cov}
\alias{canonizeRanPars} 
\alias{clean_cs} 
\alias{checkDistMatrix} 
\alias{coef.corMatern} 
\alias{coef<-.corMatern} 
\alias{corFactor.corMatern}
\alias{corMatrix.corMatern} 
\alias{corrHLfit_body}
\alias{corrMM.LRT} 
\alias{d2mudeta2fn} 
\alias{DEPARSE}
\alias{dispFn}
\alias{calc_CARdispGammaGLM} 
\alias{dispInv} 
\alias{extract.check.coords} 
\alias{findOffset} 
\alias{findSpatial}
\alias{parseBars} 
\alias{findbarsMM} 
\alias{findplus}
\alias{fullrho}
\alias{generateFileName} 
\alias{generateName} 
\alias{generateInitPhi}
\alias{getCall.HLfit} 
\alias{getCovariate.corMatern} 
\alias{getgroups} 
\alias{getPar}
\alias{getProcessed}
\alias{get_methods_ranefs}
\alias{glm.reformat}
\alias{HLCor_body}
\alias{HLfit_body}
\alias{invscalefnfn}
\alias{is.identity} 
\alias{largeLambdaMessages} 
\alias{legend_lambda}
\alias{LevenbergMstep}
\alias{LevenbergMstepCallingCpp}
\alias{locoptim}
\alias{NAcleaning}
\alias{optimthroughSmooth}
\alias{labelfn}
\alias{locatePointinvT}
\alias{logDet.corMatern} 
\alias{LogAbsDetWrap}
\alias{logL_from_SEMsample}
\alias{loglfn.ranU} 
\alias{makeCheckGeoMatrices}
\alias{makeCovEst1} 
\alias{makeCovEst2} 
\alias{makeLowerUpper}
\alias{makeTicks}
\alias{MaternCorr.ff}
\alias{Matern.corr} % back compat, should be deleted later
\alias{MLmess}
\alias{muetafn}
\alias{newetaFix}
\alias{nextPoints}
\alias{niceLabels}
\alias{nobarsMM} 
\alias{nobarsNooffset} 
\alias{noNonSpatialbarsMM}
\alias{noOffset} 
\alias{noReForm} 
\alias{nuFn} 
\alias{nuInv} 
\alias{overcat}
\alias{pastefrom} 
\alias{plotdhglm} 
\alias{plotScale}
\alias{post.process.ZALlist}
\alias{preprocess}
\alias{print.arglist}
\alias{print.predictor}
\alias{print.vcov.HLfit}
\alias{print.predictions}
\alias{print.ranef}
\alias{process_resglm_list}
\alias{RBIND}
\alias{RBIND.matrix}
\alias{RBIND.Matrix}
\alias{recalc.corMatern} 
\alias{reformat_verbose}
\alias{REMLmess}
\alias{rhoFn} 
\alias{rhoInv}
\alias{rntneg} 
\alias{rntpos} 
\alias{safesolve.qr.matrix} 
\alias{safesolve.qr.vector} 
\alias{sampleGridFromLowUp}
\alias{sampleNextPars}
\alias{scalefnfn}
\alias{selectLoglfn}
\alias{SEMbetalambda}
\alias{pMVN}
\alias{SEMdiagnosticPlot}
\alias{SEMdiagnosticPlot2D}
\alias{separator} 
\alias{setProcessed}
\alias{setControl}
\alias{Sigwrapper}
\alias{Sig_times_b}
\alias{singularSigmaMessagesStop}
\alias{slashTerms} 
\alias{selfAdjointWrapper} 
\alias{solveWrap.matrix}
\alias{solveWrap.vector}
\alias{spaMM_rhullByEI}
\alias{spaMM_bounds1D}
\alias{spaMMLRT} 
\alias{spMMFactorList} 
\alias{spMMexpandSlash}
\alias{subbarsMM} 
\alias{sweepZ1Wwrapper} 
\alias{tcrossprodWrap}
\alias{theta.mu.canonical}
\alias{theta.mu.conjugate} 
\alias{thetaMuDerivs} 
\alias{tryCatch.W.E}
\alias{u_h_v_h_from_v_h}
\alias{updateWranef}
\alias{updateW_ranefS}
\alias{validData}
\alias{ZWZtwrapper}
\alias{ZtWZwrapper}
\alias{validRangefn}
\alias{volTriangulation} %~ could become public later : cf le .Rd.true.txt
\alias{rvolTriangulation} %~ could become public later
\alias{subsimplices.volTriangulation} %~ could become public later




\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
