\name{spaMM-internal}

\alias{Bartlett.robust} 
\alias{Beta} 
\alias{Dim.corMatern} 
\alias{GammaForDispGammaGLM}
\alias{HLCor.obj} 
\alias{HL_process.args} 
\alias{HL_strictify} 
\alias{HLframes}
\alias{Initialize.corMatern} 
\alias{LR.HL} 
\alias{LRT.corrHL} 
\alias{Matern.local} 
\alias{ULI}
\alias{Variogram.corMatern} 
\alias{asStandardFormula} 
\alias{bessel_lnKnu} 
\alias{calc.XinvS}
\alias{clean_cs} 
\alias{coef.corMatern} 
\alias{coef<-.corMatern} 
\alias{corFactor.corMatern}
\alias{corMatrix.corMatern} 
\alias{corrMM.LRT} 
\alias{d2mudeta2fn} 
\alias{dispFn}
\alias{dispGammaGLM} 
\alias{dispInv} 
\alias{findOffset} 
\alias{findSpatial} 
\alias{findSpatialOrNot} 
\alias{findbarsMM} 
\alias{findplus}
\alias{formula.HLfit}
\alias{generateFileName} 
\alias{generateName} 
\alias{getCovariate.corMatern} 
\alias{locoptim}
\alias{logDet.corMatern} 
\alias{loglfn.ranU} 
\alias{make.scaled.dist} 
\alias{muetafn}
\alias{nobarsMM} 
\alias{nobarsNooffset} 
\alias{noOffset} 
\alias{nuFn} 
\alias{nuInv} 
\alias{pastefrom} 
\alias{plotdhglm} 
\alias{preprocess}
\alias{predict.HL}
\alias{print.arglist}
\alias{print.predictor}
\alias{recalc.corMatern} 
\alias{rhoFn} 
\alias{rhoInv}
\alias{safesolve.qr.matrix} 
\alias{safesolve.qr.vector} 
\alias{selectLoglfn}
\alias{separator} 
\alias{slashTerms} 
\alias{spaMMLRT} 
\alias{spMMFactorList} 
\alias{spMMexpandSlash}
\alias{subbarsMM} 
\alias{theta.mu.canonical}
\alias{theta.mu.conjugate} 
\alias{thetaMuDerivs} 
\alias{updateWranef}


\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
