\name{spaMM-internal}

%\alias{makenewname} % fixme can be removed when next (>08/2017) version of Infusion is on CRAN

\alias{fitme_body}
\alias{HLfit.obj}


%\alias{noNonSpatialbarsMM} % fixme can be removed when next (>08/2017) version of Infusion is on CRAN
%\alias{LogAbsDetWrap} % fixme can be removed when next (>08/2017) version of Infusion is on CRAN

\alias{Beta} 
\alias{spaMM_Gamma} 
\alias{get_HLCorcall} % fixme: deserves to be documented as of independent interest (cf exampl of usage in probitgem)
\alias{HLCor.obj} 
%\alias{ULI} % cf def for what to do
\alias{corrHLfit_body}
\alias{corrMM.LRT} % in Ecography code, redefined as a stop()
\alias{getCall.HLfit} 
\alias{getPar}
\alias{HLCor_body}
\alias{HLfit_body}
\alias{makeTicks} % used by probitgem and Infusion but not spaMM itself. Either I define it in each of the two other packages or I keep it here
\alias{Matern.corr} % back compat, should be deleted later
\alias{niceLabels} % used by blackbox
\alias{overcat} % in probitgem. Was in Infusion and blackbox (not in the future > 26/08/2017 versions) so cannot be removed yet.

\alias{crack} % not doc'ed but used in an example :-)


\title{Internal spaMM Functions}
\description{
  Internal spaMM functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
