\name{spaMM-S3}
\alias{spaMM-S3}

\alias{dim.precision}
\alias{print.arglist} 
\alias{print.predictor}
\alias{print.vcov.HLfit}
\alias{print.spaMM_predictions}
\alias{print.spaMM_simulations}
%\alias{print.spcall}
\alias{print.bootci4print}

% emmeans as special code in its generics to get (even S3method()-undeclared 'methods'
% from other packages).
% So the following methods are S3method()-undeclared, and may not need an \alias?
\alias{recover_data.HLfit}
\alias{emm_basis.HLfit}



\title{S3 methods of generics defined in other packages}
\description{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }