\name{spaMM-S3}
\alias{spaMM-S3}

\alias{dim.precision}
\alias{print.arglist} 
\alias{print.predictor}
\alias{print.vcov.HLfit}
\alias{print.predictions}
%\alias{print.spcall}
\alias{print.bootci4print}



\title{S3 methods of generics defined in other packages}
\description{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }